+++ CHANGELOG.md
@@ -36,6 +36,7 @@ no loot learning can occur.
 - Fix layout block changing tiers in steps of two
 - Fix duplicate texture for Scale The Summit III [#465]
 - Fix Mob Shard recipe missing for clearing in crafting table [#466]
+- Fix heart display showing wrong cell capacity [#463]
 ### Removed
 
 ## [1.16.3-1.0.1.0] - 2020-11-22